+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] ##
+
+[Unreleased]: https://github.com/mfederczuk/utest-script/compare/v1.0.2...develop
+
+### Deprecated ###
+
+* `-v` option is replaced with `-V` option
+
 ## [1.0.2] - 2019-30-01 ##
 
 [1.0.2]: https://github.com/mfederczuk/utest-script/compare/v1.0.1...v1.0.2